requirements.txt
streamlit
openai
pandas
reportlab
pyttsx3
SpeechRecognition
requests
deep-translator
gTTS
pygame
google-generativeai
pyaudio
PyPDF2
python-docx
Pillow